+++ apps/frontend/src/components/ui/ProgressBar.tsx
